requirements.txt
boto3==1.28.30
Flask==2.3.2
matplotlib==3.7.2
numpy
opencv_python==4.8.0.76
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
